Fold BreakpointManager into DebugSession
